// src/types/generated-admin-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};
// ...
export type ConfigurableOperation = {
  __typename?: 'ConfigurableOperation';
  code: Scalars['String'];
  args: Array<ConfigArg>;
};
// ...
export type CreateNfeSettingCustomInput = {
  data: NfeSettingCustomInput;
};

export type CreateNfeSettingPaymentMethodInput = {
  data: NfeSettingPaymentMethodInput;
  methodId: Scalars['ID'];
};

export type CreateNfeSettingShippingMethodInput = {
  data: NfeSettingShippingMethodInput;
  methodId: Scalars['ID'];
};
// ...
export enum LogicalOperator {
  And = 'AND',
  Or = 'OR'
}
// ...
export type MutationCreateNfeSettingPaymentMethodArgs = {
  input: CreateNfeSettingPaymentMethodInput;
};


export type MutationUpdateNfeSettingPaymentMethodArgs = {
  input: UpdateNfeSettingPaymentMethodInput;
};


export type MutationCreateNfeSettingShippingMethodArgs = {
  input: CreateNfeSettingShippingMethodInput;
};


export type MutationUpdateNfeSettingShippingMethodArgs = {
  input: UpdateNfeSettingShippingMethodInput;
};


export type MutationCreateNfeSettingCustomArgs = {
  input: CreateNfeSettingCustomInput;
};


export type MutationUpdateNfeSettingCustomArgs = {
  input: UpdateNfeSettingCustomInput;
};
// ...
export type NfeSettingCustomInput = {
  set: Array<NfeSettingCustomItemInput>;
};

export type NfeSettingCustomItem = {
  __typename?: 'NfeSettingCustomItem';
  key: Scalars['String'];
  value: Scalars['String'];
};

export type NfeSettingCustomItemInput = {
  key: Scalars['String'];
  value: Scalars['String'];
};

export type NfeSettingData = NfeSettingPaymentMethod | NfeSettingShippingMethod | NfeSettingCustom;

export type NfeSettingFilterParameter = {
  id?: Maybe<IdOperators>;
  createdAt?: Maybe<DateOperators>;
  updatedAt?: Maybe<DateOperators>;
  type?: Maybe<StringOperators>;
  methodId?: Maybe<IdOperators>;
};

export type NfeSettingList = PaginatedList & {
  __typename?: 'NfeSettingList';
  items: Array<NfeSetting>;
  totalItems: Scalars['Int'];
};

export type NfeSettingListOptions = {
  /** Skips the first n results, for use in pagination */
  skip?: Maybe<Scalars['Int']>;
  /** Takes n results, for use in pagination */
  take?: Maybe<Scalars['Int']>;
  /** Specifies which properties to sort the results by */
  sort?: Maybe<NfeSettingSortParameter>;
  /** Allows the results to be filtered */
  filter?: Maybe<NfeSettingFilterParameter>;
  /** Specifies whether multiple "filter" arguments should be combines with a logical AND or OR operation. Defaults to AND. */
  filterOperator?: Maybe<LogicalOperator>;
};

export type NfeSettingPaymentMethod = {
  __typename?: 'NfeSettingPaymentMethod';
  pagamento: Scalars['Int'];
  forma_pagamento?: Maybe<Scalars['String']>;
  tipo_integracao: Scalars['Int'];
  cnpj_credenciadora?: Maybe<Scalars['String']>;
};

export type NfeSettingPaymentMethodInput = {
  pagamento: Scalars['Int'];
  forma_pagamento?: Maybe<Scalars['String']>;
  tipo_integracao: Scalars['Int'];
  cnpj_credenciadora?: Maybe<Scalars['String']>;
};
// ...
export type NfeSettingShippingMethodInput = {
  cnpj?: Maybe<Scalars['String']>;
  razao_social?: Maybe<Scalars['String']>;
  ie?: Maybe<Scalars['String']>;
  cpf?: Maybe<Scalars['String']>;
  nome_completo?: Maybe<Scalars['String']>;
  endereco?: Maybe<Scalars['String']>;
  uf?: Maybe<Scalars['String']>;
  cidade?: Maybe<Scalars['String']>;
  cep?: Maybe<Scalars['String']>;
  placa?: Maybe<Scalars['String']>;
  uf_veiculo?: Maybe<Scalars['String']>;
  rntc?: Maybe<Scalars['String']>;
  reboque?: Maybe<NfeSettingShippingMethodReboqueInput>;
};
// ...
export type NfeSettingSortParameter = {
  id?: Maybe<SortOrder>;
  createdAt?: Maybe<SortOrder>;
  updatedAt?: Maybe<SortOrder>;
  methodId?: Maybe<SortOrder>;
};

export enum NfeSettingType {
  PaymentMethod = 'PaymentMethod',
  ShippingMethod = 'ShippingMethod',
  Custom = 'Custom'
}
// ...
export type Node = {
  id: Scalars['ID'];
};
// ...
export type PaymentMethod = Node & {
  __typename?: 'PaymentMethod';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  name: Scalars['String'];
  code: Scalars['String'];
  description: Scalars['String'];
  enabled: Scalars['Boolean'];
  checker?: Maybe<ConfigurableOperation>;
  handler: ConfigurableOperation;
  customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type QueryNfeSettingsArgs = {
  options?: Maybe<NfeSettingListOptions>;
};


export type QueryNfeSettingByTypeAndMethodIdArgs = {
  type: Scalars['String'];
  id: Scalars['ID'];
};
// ...
export type ShippingMethod = Node & {
  __typename?: 'ShippingMethod';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  languageCode: LanguageCode;
  code: Scalars['String'];
  name: Scalars['String'];
  description: Scalars['String'];
  fulfillmentHandlerCode: Scalars['String'];
  checker: ConfigurableOperation;
  calculator: ConfigurableOperation;
  translations: Array<ShippingMethodTranslation>;
  customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type ShippingMethodTranslation = {
  __typename?: 'ShippingMethodTranslation';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  languageCode: LanguageCode;
  name: Scalars['String'];
  description: Scalars['String'];
};
// ...
export type UpdateNfeSettingCustomInput = {
  id: Scalars['ID'];
  data: NfeSettingCustomInput;
};

export type UpdateNfeSettingPaymentMethodInput = {
  id: Scalars['ID'];
  data: NfeSettingPaymentMethodInput;
};